pyproject.toml
[project]
name = "freelans-bot"
version = "0.1.0"
description = "Full-auto freelance lead collector + AI proposal + Telegram notifier"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.22.0",
  "aiosqlite>=0.20.0",
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "openai>=1.109.1",
  "playwright>=1.55.0",
  "pydantic-settings>=2.11.0",
  "python-dateutil>=2.9.0.post0",
  "pyyaml>=6.0.2",
  "tenacity>=9.1.2",
  "uvicorn[standard]>=0.35.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.13.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
